requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
openai==1.3.3
requests==2.28.1
